+++ src/backup-WSL.ps1
@@ -1,4 +1,4 @@
-#---------------------------------------------------------------------------
+#--------------------------------------------------------------------------
 # PowerShell 7.5.4 
 # by Kurt Duyck (kurt.duyck@vives.be)
 #
@@ -25,7 +25,7 @@ PARAM (
 #------------------------------------------------------------------
 # START SET
   $debug = 1
-  $version = "0.3.8"
+  $version = "0.4.1"
   $app = "backup-WSL.ps1"
   $info = "backup WSL"
   $ld = "C:\tmp\log" 
@@ -56,27 +56,35 @@ function waitforenter {
 
 function check($distroparam) {
   # NEEDS rework: setup matrix: running, stopped vs known unknown
+  #             | running | stopped | known | unknown
+  # WSL F42     |         |         |       |    
+  # WSL F43     |         |         |       |    
   $wsllist = wsl --list
   $arrwsllist = $wsllist.Split("`r`n")
+  # drop the empty lines ! 
+  $newarrwsllist = $arrwsllist.Where({ $_ -ne "" }) # drop empty lines
   Write-Output "Total distros in array-->" $arrwsllist.Count
+  Write-Output "Total distros in new array-->" $newarrwsllist.Count
   # $regexdistro = $distroparam+' (Standaard)'
   $regexdistro = $distroparam
-  "-----------M-"
+  "------------"
   $regexdistro
-  "-----------M-"
-  foreach ($item in $arrwsllist) {
-    # $item
-    if ($item -match $regexdistro) {
+  "------------"
+  foreach ($item in $newarrwsllist) {
+    $item
+    if ($regexdistro -match $item) {
        Write-Host "MATCH found: $item"
     }  
-    if ($item -like $regexdistro) {
+    if ($regexdistro -like $item) {
        Write-Host "LIKE found: $item"
     }  
-    if ($item -eq $regexdistro) {
+    if ($regexdistro -eq $item) {
        Write-Host "Full match found: $item"
     } 
+    if ($regexdistro -ne $item) {
+       Write-Host "no match found: $item - $regexdistro"
+    }
   }
-  # exit
   $wsllistrunning = wsl --list --running
   $arrwsllistrunning = $wsllistrunning.Split("`r`n")
   # check if multiple distros are spinning
@@ -183,15 +191,8 @@ foreach ($arg in $args)
 #    .\$app -distro <string>
 #
 #OPTIONAL PARAMETERS
-#       -h
-#       help
-#    	-help
-#    	--help
-#       -V
-#       version
-#       -version
-#       --version
-#
+#       -h, help, -help, --help
+#       -V, version, -version, --version
 #SAMPLE
 #PS > .\$app -h
 #PS > .\$app -V
@@ -213,7 +214,6 @@ foreach ($arg in $args)
 }
 $hdata = [string]$args[0]
 write-output $hdata
-#? write-output ">"$app$version
 "---------------------------------------------------"
 write-output $info $distro
 "---------------------------------------------------"
@@ -224,4 +224,4 @@ processdata;
 prunebackups;
 #####################################################
 writelog("--------------------------------------")	
-"EOF"
+